システム変数(stat/iparam/wparam/lparam)のint64化 (Phase 2)
- HSPCTX の stat, iparam, wparam, lparam を int → int64_t に変更
- reffunc_sysvar() で stat/iparam/wparam/lparam を HSPVAR_FLAG_INT64 で返す
  → _wparam = wparam とすると _wparam は自動的にint64型になる
- SetEventNoticePtr / SetObjectEventNoticePtr を int64_t* 対応に変更
- code_adddbg() に int64_t オーバーロードを追加
- hsp3, hsp3net 32bit/64bit 全ビルド確認済み
- テスト: 基本演算・stat取得が正常動作

+++ hsp3/hsp3struct.h
@@ -543,9 +543,9 @@ struct HSPCTX
 
 	IRQDAT *mem_irq;					// IRQ data ptr
 	int irqmax;							// IRQ data count
-	int iparam;							// IRQ Info data1
-	int wparam;							// IRQ Info data2
-	int lparam;							// IRQ Info data3
+	int64_t iparam;						// IRQ Info data1
+	int64_t wparam;						// IRQ Info data2
+	int64_t lparam;						// IRQ Info data3
 
 	PVal *mem_var;						// var storage index
 	HSPEXINFO30 exinfo;					// HSP function data(3.0)
@@ -559,7 +559,7 @@ struct HSPCTX
 	int looplev;						// repeat loop level
 	HSPERROR err;						// error code
 	int hspstat;						// HSP status
-	int stat;							// sysvar 'stat'
+	int64_t stat;						// sysvar 'stat'
 	int strsize;						// sysvar 'strsize'
 	char *refstr;						// RefStr Buffer
 	char *fnbuffer;						// buffer for FILENAME
