+++ hsp3net/win32gui/hspwnd_win.h
@@ -331,7 +331,7 @@ public:
 	int GetActive( void );
 	void SetNotifyFunc( void *func );
 	int GetBmscrMax( void ) { return bmscr_max; };
-	void SetEventNoticePtr( int *ptr );
+	void SetEventNoticePtr( int64_t *ptr );
 	void SetParentWindow( void *hwnd ) { wnd_parent = hwnd; };
 	int GetEmptyBufferId(void);
 	int GetPreloadBufferId(char *fname);
@@ -354,9 +354,9 @@ private:
 	Bmscr **mem_bm;
 	WNDPROC* wfWndProc;
 	int bmscr_max;
-	int bmscr_res;
+	int64_t bmscr_res;
 	int wfx,wfy,wbx,wby;
-	int *resptr;
+	int64_t *resptr;
 	TCHAR defcls[32];			// Default Window Class
 	void *wnd_parent;			// Parent Window Handle
 };
@@ -452,7 +452,7 @@ typedef struct BMSCR
 	std::string resname;				// Resource Name
 } BMSCR;
 
-void SetObjectEventNoticePtr( int *ptr );
+void SetObjectEventNoticePtr( int64_t *ptr );
 
 
 #endif
